==> scMultiome/GEX_analysis/03.2_neuronal_DEA.Rmd <==
---
title: "RNA neuronal DEA"
output: html_document
date: "2024-11-11"
---

```{r}
suppressPackageStartupMessages({
  library(SingleCellExperiment)
  library(sechm)
  library(ggplot2)
  library(patchwork)
  library(edgeR)
  library(scuttle)
  library(Matrix)
  library(muscat)
  library(future)
  library(BiocParallel)
})
theme_set(theme_bw())
source("functions.R")
```

```{r}
pb <- readRDS("neuronal.celltype.PB.rds")

res <- bplapply(split(seq_len(ncol(pb)), pb$celltype), BPPARAM=MulticoreParam(17, progress=FALSE), FUN=function(x){
  se <- pb[,x]
  if(median(se$ncells)>50){ 
    set.seed(123)
    se <- SEtools::svacor(se, ~TimePoint*Treatment, n.sv=1)
    mm <- model.matrix(~SV1+TimePoint*Treatment, data=as.data.frame(colData(se)))
    coRm <- -1:-2
  }else{
    se$SV1 <- NA_real_
    assays(se)$corrected <- assays(se)$logcpm
    mm <- model.matrix(~TimePoint*Treatment, data=as.data.frame(colData(se)))
    coRm <- -1
  }
  se <- log2FC(se, "corrected", se$TimePoint=="Control" & se$Treatment=="Handling", isLog = TRUE)
  dds <- calcNormFactors(DGEList(assay(se)), method="TMMwsp")
  dds <- dds[filterByExpr(dds,mm,min.count=20),]
  dds <- estimateDisp(dds,mm)
  fit <- glmFit(dds,mm)
  comps <- c(list(stressAny=grep("TimePoint",colnames(mm)),
                treatAny=grep("Treatment",colnames(mm))),
           lapply(setNames(seq_len(ncol(mm))[coRm],colnames(mm)[coRm]), FUN=identity))
  deas <- lapply(comps, FUN=function(x){
    as.data.frame(topTags(glmLRT(fit, coef=x), Inf))
  })
  list(se=se, deas=deas)
})
pb <- do.call(cbind, lapply(res, FUN=function(x) x$se))

# produce a muscat-like table
names(comps) <- comps <- names(res[[1]]$deas)
res1 <- list(table=lapply(comps, FUN=function(co){
  lapply(setNames(names(res),names(res)), FUN=function(ctn){
    ct <- res[[ctn]]$deas[[co]]
    ct <- ct[which(ct$logCPM>=1.5),]
    ct$gene <- row.names(ct)
    ct$p_adj.loc <- ct$FDR
    ct$p_val <- ct$PValue
    ct$cluster_id <- as.factor(ctn)
    row.names(ct) <- NULL
    ct
  })
}))
names(res1) <- gsub("TimePoint","ARS",names(res1))

# bulk FDR as covariate
ud <- readRDS("../../../work/plger/CRS/rnaDEAs/all.unspliced.SE.rds")
fd <- readRDS("../../../work/plger/CRS/rnaDEAs/all.full.SE.rds")
bulkfdr <- setNames(rowMins(cbind(rowData(fd)$FDR,rowData(ud)$FDR),na.rm=TRUE),row.names(ud))
bulkfdr[is.infinite(bulkfdr) | is.na(bulkfdr)] <- 2

res <- applyIHWonMuscatRes(res=res1, pb=pb, bulkfdr = bulkfdr)
pb <- res$pb
res <- res$res

ndegs <- sapply(res$table, FUN=function(x){
  sapply(x, FUN=function(y){
    sum(y$ihw<0.05,na.rm=TRUE)
  })
})

res2 <- dplyr::bind_rows(lapply(res$table, FUN=dplyr::bind_rows), .id="Comparison")
for(f in c("Comparison","cluster_id")) res2[[f]] <- factor(res2[[f]])
res2$LR <- res2$PValue <- res2$FDR <- res2$coef <- NULL
saveRDS(res2, "neuronal.DEA.rds")

metadata(pb)$anno_colors <- list(TimePoint=setNames(RColorBrewer::brewer.pal(4,"Reds"), levels(pb$TimePoint)),
                                 Treatment=c(Handling="lightgrey", Restraint="midnightblue"))
metadata(pb)$default_view <- list(top_annotation=c("TimePoint","Treatment"))
pb$celltype <- factor(as.character(pb$celltype), sort(levels(pb$celltype)))
pb$celltype2 <- gsub("/","\n",pb$celltype)
pb$celltype2 <- factor(as.character(pb$celltype2), sort(unique(pb$celltype2)))
saveRDS(pb, "neuronal.celltype.PB.rds")
```

```{r}
# specific ihw for baseline restraint:
pb <- readRDS("neuronal.celltype.PB.rds")
dea1 <- getDEA(ud, "CRS.baseline")
dea2 <- getDEA(fd, "CRS.baseline")
dea2$FDR2 <- dea1[row.names(dea2),"FDR"]
dea2 <- dea2[!is.na(dea2$logCPM),]
dea2$minFDR <- pmin(dea2$FDR, dea2$FDR2, na.rm=TRUE)
bulkfdr <- setNames(dea2$minFDR, row.names(dea2))
res2 <- readRDS("neuronal.DEA.rds")
res2 <- res2[res2$Comparison=="TreatmentRestraint",]
res2$bulkFDR <- bulkfdr[res2$gene]
res2$bulkFDR[is.infinite(res2$bulkFDR) | is.na(res2$bulkFDR)] <- 2
a <- IHW::ihw(res2$p_val, res2$bulkFDR, alpha=0.1)
res2$ihw <- IHW::adj_pvalues(a)
head(res2[order(res2$ihw, res2$p_val),])
resR <- res2
res2 <- readRDS("neuronal.DEA.rds")
res2 <- res2[res2$Comparison!="TreatmentRestraint",]
res2 <- rbind(res2,resR)
saveRDS(res2, file="neuronal.DEA.rds")
```


```{r, fig.width=14, fig.height=8}
res3 <- res2[!grepl("Any|Restraint",res2$Comparison),]
tp <- c("15min","45min","3h")
res3$Comparison <- factor(as.character(res3$Comparison), paste0("TimePoint",tp), paste0("ARS",tp))
celltypeVolcanos(res3)
```


```{r, fig.width=11, fig.height=6}
degs <- unique(res3$gene[res3$ihw<0.25])
sechm(pb, degs, gaps_at="celltype", top_annotation = c("Treatment","TimePoint"))
```


```{r, fig.width=13, fig.height=5}
args <- strsplit("Gadd45g, Egr4, Arc, Fosb, Npas4, Btg2, Fos, Junb, Jun, Egr1, Dusp5, Egr3, Nr4a1, Nr4a2, Per1, Scg2",", ")[[1]]
sechm(pb, args, gaps_at="celltype", top_annotation = c("Treatment","TimePoint"))
res3$celltype <- gsub("\\.|/","\n",res3$cluster_id)
res5 <- res3[res3$gene %in% args,]
ggplot(res5, aes(celltype, logFC, fill=Comparison)) + geom_hline(yintercept=0, linetype="dashed") +
  geom_boxplot(outlier.colour = NA) + coord_cartesian(ylim=c(-2.5,5)) + theme_bw() +
  labs(x="", y="logFC of selected ARGs") + theme(legend.position = "bottom")
```


```{r, fig.width=14, fig.height=6}
g <- strsplit("Sbk1, Prkab2, Mrps6, Lefty1, Rem2, Hmgcr, Kpna4, Nr4a3, Nr4a1, Nr4a2, Manf, Rasgrp1",", ")[[1]]
sechm(pb[,pb$celltype %in% names(res$table[[1]])], args, gaps_at="celltype2", top_annotation = c("TimePoint","Treatment","ncells"))
sechm(pb[,pb$celltype %in% names(res$table[[1]])], g, gaps_at="celltype2", top_annotation = c("TimePoint","Treatment","ncells"))
```

# Broad classes


```{r}
sce <- readRDS("neuronal.SCE.rds")
w <- grep("^ExN",sce$celltype)
pb <- sumCountsAcrossCells(sce[,w],
                           colData(sce)[w,c("Sample","Treatment","TimePoint","GEM_batch")])
colnames(pb) <- pb$Sample
pb <- pb[,order(pb$TimePoint, pb$Treatment)]
dds <- calcNormFactors(DGEList(assay(pb)))
assays(pb)$logcpm <- log1p(edgeR::cpm(dds))
pb <- SEtools::svacor(pb, ~Treatment*TimePoint, n.sv=1)
pb <- log2FC(pb, "corrected", pb$TimePoint=="Control" & pb$Treatment=="Handling", isLog = TRUE)
mm <- model.matrix(~SV1+Treatment*TimePoint, data=as.data.frame(colData(pb)))
dds <- estimateDisp(dds,mm)
dds <- dds[filterByExpr(dds,mm,min.count=15),]
fit <- glmFit(dds,mm)
for(f in colnames(mm)[-1:-2]){
  message(f)
  res <- as.data.frame(topTags(glmLRT(fit, f),Inf))
  res$bulkFDR <- bulkfdr[row.names(res)]
  res$bulkFDR[is.na(res$bulkFDR)] <- 2
  a <- IHW::ihw(res$PValue, res$FDR, alpha=0.1)
  res$ihw <- IHW::adj_pvalues(a)
  rowData(pb)[[paste0("DEA.",gsub("Treatment|TimePoint","",f))]] <- res[row.names(pb),]
}
saveRDS(pb, "neuronal.excitatory.PB.rds")

w <- grep("^InhN",sce$celltype)
pb <- sumCountsAcrossCells(sce[,w],
                           colData(sce)[w,c("Sample","Treatment","TimePoint","GEM_batch")])
colnames(pb) <- pb$Sample
pb <- pb[,order(pb$TimePoint, pb$Treatment)]
dds <- calcNormFactors(DGEList(assay(pb)))
assays(pb)$logcpm <- log1p(edgeR::cpm(dds))
pb <- SEtools::svacor(pb, ~Treatment*TimePoint, n.sv=1)
pb <- log2FC(pb, "corrected", pb$TimePoint=="Control" & pb$Treatment=="Handling", isLog = TRUE)
mm <- model.matrix(~SV1+Treatment*TimePoint, data=as.data.frame(colData(pb)))
dds <- estimateDisp(dds,mm)
dds <- dds[filterByExpr(dds,mm),]
fit <- glmFit(dds,mm)
for(f in colnames(mm)[-1:-2]){
  message(f)
  res <- as.data.frame(topTags(glmLRT(fit, f),Inf))
  res$bulkFDR <- bulkfdr[row.names(res)]
  res$bulkFDR[is.na(res$bulkFDR)] <- 2
  a <- IHW::ihw(res$PValue, res$FDR, alpha=0.1)
  res$ihw <- IHW::adj_pvalues(a)
  rowData(pb)[[paste0("DEA.",gsub("Treatment|TimePoint","",f))]] <- res[row.names(pb),]
}
saveRDS(pb, "neuronal.inhibitory.PB.rds")
```


```{r}
sessionInfo()
```
